+++ contracts/0.8.25/vaults/interfaces/IBeaconProxy.sol
@@ -6,6 +6,5 @@ pragma solidity 0.8.25;
 
 interface IBeaconProxy {
     function getBeacon() external view returns (address);
-
-    function version() external pure returns(uint8);
+    function version() external pure returns(uint256);
 }
